pyproject.toml
[project]
name = "gtfo-mapgeneration"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "keyboard>=0.13.5",
    "lxml>=6.0.2",
    "numpy>=2.3.5",
    "pillow>=12.0.0",
    "shapely>=2.1.2",
]
